+++ src/ninja/mbedded/ninjaterm/model/terminal/txRx/TxRx.java
@@ -1,8 +1,6 @@
 package ninja.mbedded.ninjaterm.model.terminal.txRx;
 
-import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.property.SimpleStringProperty;
-import javafx.scene.input.KeyEvent;
 import ninja.mbedded.ninjaterm.model.terminal.txRx.layout.Layout;
 
 /**
@@ -12,15 +10,29 @@ public class TxRx {
 
     public Layout layout = new Layout();
 
+
+    public SimpleStringProperty txData = new SimpleStringProperty("");
+
     /**
      * Initialise with "" so that it does not display "null".
      */
-    public SimpleStringProperty txRxText = new SimpleStringProperty("");
+    public SimpleStringProperty txRxData = new SimpleStringProperty("");
 
+    public void addTxData(String data) {
+
+        txData.set(txData.get() + data);
+
+        // Echo TX data into TX/RX pane
+        if(layout.localTxEcho.get()) {
+            txRxData.set(txRxData.get() + data);
+        }
+
+    }
 
     public void addRxData(String data) {
-        txRxText.set(txRxText.get() + data);
+        txRxData.set(txRxData.get() + data);
     }
+
 
 
 }